Made support for legacy lifecycle hook annotations automatic
