+++ elide-core/src/main/java/com/yahoo/elide/core/dictionary/EntityDictionary.java
@@ -16,6 +16,18 @@ import com.yahoo.elide.annotation.Exclude;
 import com.yahoo.elide.annotation.Include;
 import com.yahoo.elide.annotation.LifeCycleHookBinding;
 import com.yahoo.elide.annotation.NonTransferable;
+import com.yahoo.elide.annotation.OnCreatePostCommit;
+import com.yahoo.elide.annotation.OnCreatePreCommit;
+import com.yahoo.elide.annotation.OnCreatePreSecurity;
+import com.yahoo.elide.annotation.OnDeletePostCommit;
+import com.yahoo.elide.annotation.OnDeletePreCommit;
+import com.yahoo.elide.annotation.OnDeletePreSecurity;
+import com.yahoo.elide.annotation.OnReadPostCommit;
+import com.yahoo.elide.annotation.OnReadPreCommit;
+import com.yahoo.elide.annotation.OnReadPreSecurity;
+import com.yahoo.elide.annotation.OnUpdatePostCommit;
+import com.yahoo.elide.annotation.OnUpdatePreCommit;
+import com.yahoo.elide.annotation.OnUpdatePreSecurity;
 import com.yahoo.elide.annotation.SecurityCheck;
 import com.yahoo.elide.core.PersistentResource;
 import com.yahoo.elide.core.RequestScope;
@@ -40,6 +52,7 @@ import com.yahoo.elide.core.utils.ClassScanner;
 import com.yahoo.elide.core.utils.coerce.CoerceUtil;
 import com.yahoo.elide.core.utils.coerce.converters.Serde;
 import com.google.common.base.Preconditions;
+import com.google.common.base.Throwables;
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 import com.google.common.collect.Maps;
@@ -962,13 +975,15 @@ public class EntityDictionary {
 
         bindJsonApiToEntity.put(Pair.of(type, version), declaredClass);
         apiVersions.add(version);
-        entityBindings.put(declaredClass, new EntityBinding(this, declaredClass,
-                type, version, hiddenAnnotations));
+        EntityBinding binding = new EntityBinding(this, declaredClass, type, version, hiddenAnnotations);
+        entityBindings.put(declaredClass, binding);
 
         Include include = (Include) getFirstAnnotation(declaredClass, Arrays.asList(Include.class));
         if (include.rootLevel()) {
             bindEntityRoots.add(declaredClass);
         }
+
+        bindLegacyHooks(binding);
     }
 
     /**
@@ -1879,5 +1894,107 @@ public class EntityDictionary {
         return object instanceof Dynamic
                 ? ((Dynamic) object).getType()
                 : new ClassType(object.getClass());
+    }
+
+    public void bindLegacyHooks(EntityBinding binding) {
+        binding.getAllMethods().stream()
+                .forEach(method -> {
+                    String annotationField = null;
+                    LifeCycleHookBinding.TransactionPhase phase = null;
+                    LifeCycleHookBinding.Operation operation = null;
+                    if (method.isAnnotationPresent(OnCreatePostCommit.class)) {
+                        annotationField = method.getAnnotation(OnCreatePostCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.POSTCOMMIT;
+                        operation = LifeCycleHookBinding.Operation.CREATE;
+                    } else if (method.isAnnotationPresent(OnCreatePreCommit.class)) {
+                        annotationField = method.getAnnotation(OnCreatePreCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRECOMMIT;
+                        operation = LifeCycleHookBinding.Operation.CREATE;
+                    } else if (method.isAnnotationPresent(OnCreatePreSecurity.class)) {
+                        annotationField = method.getAnnotation(OnCreatePreSecurity.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRESECURITY;
+                        operation = LifeCycleHookBinding.Operation.CREATE;
+                    } else if (method.isAnnotationPresent(OnUpdatePostCommit.class)) {
+                        annotationField = method.getAnnotation(OnUpdatePostCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.POSTCOMMIT;
+                        operation = LifeCycleHookBinding.Operation.UPDATE;
+                    } else if (method.isAnnotationPresent(OnUpdatePreCommit.class)) {
+                        annotationField = method.getAnnotation(OnUpdatePreCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRECOMMIT;
+                        operation = LifeCycleHookBinding.Operation.UPDATE;
+                    } else if (method.isAnnotationPresent(OnUpdatePreSecurity.class)) {
+                        annotationField = method.getAnnotation(OnUpdatePreSecurity.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRESECURITY;
+                        operation = LifeCycleHookBinding.Operation.UPDATE;
+                    } else if (method.isAnnotationPresent(OnReadPostCommit.class)) {
+                        annotationField = method.getAnnotation(OnReadPostCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.POSTCOMMIT;
+                        operation = LifeCycleHookBinding.Operation.READ;
+                    } else if (method.isAnnotationPresent(OnReadPreCommit.class)) {
+                        annotationField = method.getAnnotation(OnReadPreCommit.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRECOMMIT;
+                        operation = LifeCycleHookBinding.Operation.READ;
+                    } else if (method.isAnnotationPresent(OnReadPreSecurity.class)) {
+                        annotationField = method.getAnnotation(OnReadPreSecurity.class).value();
+                        phase = LifeCycleHookBinding.TransactionPhase.PRESECURITY;
+                        operation = LifeCycleHookBinding.Operation.READ;
+                    } else if (method.isAnnotationPresent(OnDeletePostCommit.class)) {
+                        phase = LifeCycleHookBinding.TransactionPhase.POSTCOMMIT;
+                        operation = LifeCycleHookBinding.Operation.DELETE;
+                    } else if (method.isAnnotationPresent(OnDeletePreCommit.class)) {
+                        phase = LifeCycleHookBinding.TransactionPhase.PRECOMMIT;
+                        operation = LifeCycleHookBinding.Operation.DELETE;
+                    } else if (method.isAnnotationPresent(OnDeletePreSecurity.class)) {
+                        phase = LifeCycleHookBinding.TransactionPhase.PRESECURITY;
+                        operation = LifeCycleHookBinding.Operation.DELETE;
+                    }
+
+                    if (operation != null) {
+                        if (annotationField != null && ! annotationField.isEmpty()) {
+                            if (annotationField.equals("*")) {
+                                bindTrigger(binding.entityClass, operation, phase,
+                                        generateHook((Method) method), true);
+                            } else {
+                                bindTrigger(binding.entityClass, annotationField, operation, phase,
+                                        generateHook((Method) method));
+                            }
+                        } else {
+                            bindTrigger(binding.entityClass, operation, phase,
+                                    generateHook((Method) method), false);
+                        }
+                    }
+
+                });
+    }
+
+    private static LifeCycleHook generateHook(Method method) {
+        return new LifeCycleHook() {
+            @Override
+            public void execute(LifeCycleHookBinding.Operation operation,
+                                LifeCycleHookBinding.TransactionPhase phase,
+                                Object model,
+                                com.yahoo.elide.core.security.RequestScope scope,
+                                Optional changes) {
+                try {
+                    int paramCount = method.getParameterCount();
+                    Class<?>[] paramTypes = method.getParameterTypes();
+
+                    if (changes.isPresent() && paramCount == 2
+                            && paramTypes[0].isInstance(scope)
+                            && paramTypes[1].isInstance(changes.get())) {
+                        method.invoke(model, scope, changes.get());
+                    } else if (paramCount == 1 && paramTypes[0].isInstance(scope)) {
+                        method.invoke(model, scope);
+                    } else if (paramCount == 0) {
+                        method.invoke(model);
+                    } else {
+                        throw new IllegalArgumentException();
+                    }
+                } catch (ReflectiveOperationException e) {
+                    Throwables.propagateIfPossible(e.getCause());
+                    throw new IllegalArgumentException(e);
+                }
+            }
+        };
     }
 }
